+++ Disciple/Changes.md
@@ -6,6 +6,8 @@
 ### New features
 - **Gateway** now supports custom redirect strategies
 - **RequestRateLimiter** now supports temporary lockdowns using `.lockUntil(Instant)` and `.removeLocking()`
+### Other changes
+- **RequestRateLimiter** now extends **ScopeUsable**
 
 ## v1.9.2 - 15.03.2026
 A minor update adding some quality-of-life changes, and support for some **Echo**-specific use-cases.